+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.9.11]
+- Security: Fixed to properly handle case of multiple `cfnetwork` interfaces per device
+    > Note: now dst IP check is enforced on interface & DNAT level
+
 ## [0.9.10]
 - Fixed minor Puppet Lanaguage issue appeared with 4.6.0: PUP-6606
 
@@ -63,6 +67,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release
 
+[0.9.11](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.11)
 [0.9.10](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.10)
 [0.9.9](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.9)
 [0.9.8](https://github.com/codingfuture/puppet-cffirehol/releases/tag/v0.9.8)